requirements.txt
streamlit
langchain_google_genai
langchain_core
PyYAML